+++ src/components/Menus/Sidebar/DownloadPanel/DownloadPanel.jsx
@@ -2,6 +2,8 @@ import React from "react";
 import SidebarPanel from "../SidebarPanel";
 import { useTranslation } from "react-i18next";
 import Button from "../../../Button/Button";
+import WebComponentDownloadButton from "../../../WebComponentDownloadButton/WebComponentDownloadButton";
+import DownloadIcon from "../../../../assets/icons/download.svg";
 import "../../../../assets/stylesheets/DownloadPanel.scss";
 
 export const DownloadPanel = () => {
@@ -25,6 +27,11 @@ export const DownloadPanel = () => {
           />
         </div>
         <p>{t("downloadPanel.downloadHint")}</p>
+        <WebComponentDownloadButton
+          buttonText={t("downloadPanel.downloadButton")}
+          className="btn btn--secondary download-panel__button"
+          Icon={DownloadIcon}
+        />
       </div>
     </SidebarPanel>
   );