+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Added ability to enable sidebar in web component (#738)
 - Added ability to customise sidebar options (#738)
+- Added download panel for sidebar (#744)
 
 ### Changed